# api-cms/docs/swagger.yaml
definitions:
...
  admin.Source:
    properties:
      avatar:
        type: string
      record:
        $ref: '#/definitions/admin.SourceRecord'
      typeSource:
        type: string
      url_start:
        type: string
    type: object
  admin.SourceRecord:
    properties:
      blocked:
        type: boolean
    type: object
  auth.Account:
    properties:
      password:
        type: string
      remember:
        type: boolean
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.LoginResponse:
    properties:
      code:
        type: integer
      data:
        properties:
          access_token:
            type: string
          expire:
            type: string
        type: object
      message:
        type: string
      version:
        type: string
    type: object
  auth.RegisterAccount:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Đăng nhập hệ thôống CMS, lấy về access_token
      parameters:
      - description: username password
        in: body
        name: LoginForm
        required: true
        schema:
          $ref: '#/definitions/auth.Account'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
      summary: Đăng nhập CMS
      tags:
      - user
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Tạo tài khoản mới
      parameters:
      - description: username password
        in: body
        name: RegisterForm
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
      summary: Đăng ký tài khoản
      tags:
      - user
...
  /api/v1/common/sources:
    get:
      consumes:
      - application/json
      description: Xem danh sách Source
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem danh sách Source
      tags:
      - Source
    post:
      consumes:
      - application/json
      description: Tạo mới Source
      parameters:
      - description: name
        in: body
        name: AccountForm
        required: true
        schema:
          $ref: '#/definitions/admin.Source'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Tạo mới Source
      tags:
      - Source
  /api/v1/common/sources/{source_id}:
    delete:
      consumes:
      - application/json
      description: Xóa Source
      parameters:
      - description: source_id
        in: path
        name: source_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xóa Source
      tags:
      - Source
    get:
      consumes:
      - application/json
      description: Xem chi tiết Source
      parameters:
      - description: source_id
        in: path
        name: source_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem chi tiết Source
      tags:
      - Source
    put:
      consumes:
      - application/json
      description: Cập nhật Source
      parameters:
      - description: ID of the SOurce
        in: path
        name: source_id
        required: true
        type: string
      - description: Thông tin cập nhật của Source
        in: body
        name: SourceForm
        required: true
        schema:
          $ref: '#/definitions/admin.Source'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Cập nhật Source
      tags:
      - Source
  /api/v1/common/sources/type:
    get:
      consumes:
      - application/json
      description: Xem danh sách Source theo nguồn
      parameters:
      - description: Type Source
        in: query
        name: typeSource
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem danh sách Source theo nguồn
      tags:
      - Source
swagger: "2.0"